requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
gunicorn==21.2.0
python-dotenv==1.0.0
requests
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
sqlalchemy
schedule
aiofiles
aiohttp
librosa
numpy
opencv-python
pyyaml
torch
werkzeug